pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recipes"
version = "0.1.0"
description = "A recipe management app with meal planning and pantry tracking"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "httpx",
    "beautifulsoup4",
    "anthropic",
    "click",
    "pytest",
    "pytest-asyncio",
    "python-dotenv>=1.2.2",
]

[project.scripts]
recipes = "recipes.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/recipes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
